uric_api/requirements.txt
aioredis==1.3.1
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
bcrypt==4.2.0
billiard==4.2.0
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
channels==2.3.1
channels-redis==2.4.1
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cron-descriptor==1.4.5
cryptography==43.0.1
daphne==2.5.0
Django==5.1.1
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-cors-headers==4.4.0
django-redis==5.4.0
django-simpleui==2024.8.28
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
et-xmlfile==1.1.0
hiredis==3.0.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
kombu==5.4.1
msgpack==0.6.2
multi_key_dict==2.0.3
mysql==0.0.3
mysqlclient==2.2.4
openpyxl==3.1.5
paramiko==3.5.0
pbr==6.1.0
pillow==10.4.0
pip==24.2
prompt_toolkit==3.0.47
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==1.7.1
PyMySQL==1.1.0
PyNaCl==1.5.0
pyOpenSSL==24.2.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-gitlab==4.11.1
python-jenkins==1.8.2
pytz==2024.2
redis==5.0.8
requests==2.32.3
requests-toolbelt==1.0.0
service-identity==24.1.0
setuptools==73.0.1
six==1.16.0
sqlparse==0.5.1
tomli==2.0.1
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
wheel==0.44.0
zope.interface==7.0.3
